+++ src/components/ChapterEmoji/chapteremoji.css
@@ -1,0 +1,104 @@
+.section3 {
+    display: flex;
+    justify-content: center;
+    gap: 5%;
+    margin-top: 3%;
+    margin-bottom: 3%;
+}
+
+.emojis {
+    width: 8%;
+    height: 8%;
+}
+
+.section4 {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.rectangle-manga {
+    width: 50%;
+    height: 7rem;
+    margin-top: 3%;
+    margin-bottom: 5%;
+    background: #FFFFFF;
+    box-shadow: 0px 2px 7px rgba(0, 0, 0, 0.15);
+    border-radius: 20px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 1.5rem;
+
+}
+
+.rectangle-text {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+
+.text1 {
+    font-family: 'Montserrat';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 29px;
+    text-align: center;
+    color: #424242;
+    margin-bottom: 8px;
+}
+
+.text2 {
+    font-family: 'Montserrat';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 12px;
+    text-align: center;
+    color: #9D9D9D;
+}
+
+.text3 {
+    color: #9D9D9D;
+}
+
+@media (min-width:320px) and (max-width:768px) {
+    .section3 {
+        margin-bottom: 4%;
+        margin-top: 4%;
+    }
+
+    .emojis {
+        width: 18%;
+    }
+
+    .rectangle-manga {
+        width: 93%;
+        height: 5rem;
+        gap: 0;
+        margin-bottom: 6%;
+    }
+
+    .rectangle-text {
+        text-align: center;
+    }
+}
+
+@media (min-width:320px) and (max-width:430px) {
+    .text1 {
+        font-weight: 400;
+        font-size: 20px;
+        color: #424242;
+        margin-bottom: 0px;
+    }
+
+    .text2 {
+        font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: 400;
+        font-size: 12px;
+        color: #9D9D9D;
+    }
+}
