+++ src/components/ChapterRead/chapterread.css
@@ -1,0 +1,118 @@
+.details-btns {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 30px;
+    border-radius: 50000px;
+    border: none;
+    box-shadow: 0px 0px 7px rgba(0, 0, 0, 0.15);
+    width: 280px;
+    margin: 0 auto;
+    margin-bottom: 20px;
+
+
+}
+
+
+.manga-btn {
+    width: 80px;
+    height: 30px;
+    background: #EBEBEB;
+    border-radius: 50000px;
+    border: none;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-weight: 700;
+    font-size: 14px;
+    line-height: 17px;
+    color: #9D9D9D;
+    letter-spacing: 0.05em;
+    text-decoration: none;
+    margin: 0 12px;
+}
+
+.prueba {
+    background: linear-gradient(153.39deg, #F9A8D4 -13.9%, #F472B6 58.69%);
+    color: #FFFFFF;
+    margin: 0px;
+    width: 200px;
+}
+
+.description-manga p {
+    text-align: center;
+    margin-bottom: 10px;
+}
+
+.sectionChapter {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    gap: 5%;
+    margin: 20px 0px 20px 0px;
+}
+
+.selecChapter {
+    width: 15%;
+    height: 6%;
+    border-radius: 8px;
+}
+
+.order-chapter {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 15px;
+    font-size: 24px;
+    text-align: center;
+}
+
+.coment-chapter {
+    display: flex;
+    justify-content: center;
+    gap: 8px;
+}
+
+.coment-chapter img {
+    width: 28px;
+    height: 28px;
+}
+
+.btn-read {
+    width: 151px;
+    height: 74px;
+    background: linear-gradient(153.39deg, #F9A8D4 -13.9%, #F472B6 58.69%);
+    border-radius: 50000px;
+    border: none;
+    color: #FFFFFF;
+    font-family: 'Montserrat';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 24px;
+
+}
+
+@media (min-width:320px) and (max-width:768px) { 
+    .selecChapter {
+        width: 20%;
+    }
+    .p-chapter{
+        font-size: 16px;
+    }
+    .btn-read {
+        width: 120px;
+        height: 60px;   
+        font-weight: 500;
+        font-size: 20px;
+    }
+    .order-chapter {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        gap: 15px;
+        font-size: 18px;
+    }
+}
